feat: Add aperture (and permeability) update on fracture opening
* aperture is the combined effect of initial aperture,
which is the hydraulic aperture from the given initial permeability,
and the mechanical aperture from fracture opening as computed during
simulations.
* Compute specific volume and permeability from this aperture
* Re-compute all parameters every Newton iteration to account for the
updated aperture, which affects the specific volume of scalar terms in
fractures, and the permeability of the fractures.
* NOTE: In the next commit, probably, adjustments to fracture pressure
will be introduced to account for the changing volume, after two
proposed methods have been tested.
* in flow.py: Explicitly scale permeability
* in flow.py: Export aperture
* in contact_mechanics_biot.py: Remove after_newton_iteration
since it is identical to parent class.
* in isc_model.py and parameter.py: Compute aperture from initial
permeability in FlowParameters class, then the ISCModel class always
computes permeability from aperture (cubic law) in fractures. In 3d,
permeability should be set in pp.Rock
* Optionally compute specific volume from iterate (since you can
compute aperture from iterate).