fix: increase tolerance in InjectionRate and TimeStep Phases
This solves in issue in degenerate cases where if you have a very large
time step, then X + tol = X.
I.e., the tolerance is too small to be added to the large number. This
causes problems on phase boundaries.
This solution assumes that time steps are quite a bit larger than
1 second